Added Ant animations (partialy completed) Made some quest scripts (item, enemykill) Added rusty hatchet

+++ Assets/Scripts/PlayerStats.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class PlayerStats : MonoBehaviour {
 
     public int currentLevel;
-    public int currentExp;
+    public float currentExp;
 
     public int[] toLevelUp;
